smartclide-template-code-generation/requirements.txt
setuptools
argparse
pyyaml
python-Levenshtein